% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_features.R
\name{Adjust_Feature_Type}
\alias{Adjust_Feature_Type}
\title{Adjust_Feature_Type}
\usage{
Adjust_Feature_Type(X, var.type.vec, ordered.cat.levels.vec)
}
\arguments{
\item{X}{n x p dataframe object (allows for mixed data types) or a matrix object (single data type)}

\item{var.type.vec}{p x 1 vector with elements (values of 0-3) indicating the intended data type of each column of data.x. Mode numeric (0) will be considered as interval scaled variables; mode character or class factor (1) will be considered as categorical nominal variables; class ordered (2) will be considered as ordered categorical variables; mode logical (3) will be considered as binary asymmetric variables (view details of gower.dist() for more info).}

\item{ordered.cat.levels.vec}{(Optional) vector is required if columns of type 2 (ordered categorical variables) are specified. Say there are e variables of type 2 specified in var.type.vec and i=1,...,e. The ith element of ordered.cat.levels.vec is a list of "ordered levels" corresponding to the ith ordered categorical variable (i.e. the number of elements in ordered.cat.levels.vec should match the number of ordered categorical variables specified in var.type.vec). This is necessary for setting features to this feature type.}
}
\value{
X
}
\description{
Adjust_Feature_Type
}
\keyword{internal}
