% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_checks.R
\name{Cluster_Vis_Check}
\alias{Cluster_Vis_Check}
\title{Cluster_Vis_Check}
\usage{
Cluster_Vis_Check(cluster.vis)
}
\arguments{
\item{cluster.vis}{Boolean object, "TRUE" indicating dendrogram plot and silhouette plot are a desired output, "FALSE" otherwise.}
}
\value{
Returns an error message if cluster.vis is not a logical type object.
}
\description{
Cluster_Vis_Check
}
\keyword{internal}
