% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_checks.R
\name{Dend_Check}
\alias{Dend_Check}
\title{Dend_Check}
\usage{
Dend_Check(dend)
}
\arguments{
\item{dend}{}
}
\value{
Returns an error message if dend is not a logical type object.
}
\description{
Dend_Check
}
\keyword{internal}
