% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_checks.R
\name{Gower_Mat_Check}
\alias{Gower_Mat_Check}
\title{Gower_Mat_Check}
\usage{
Gower_Mat_Check(X)
}
\arguments{
\item{X}{n x p dataframe object (allows for mixed data types) or a matrix object (single data type)}
}
\value{
Returns an error message if X is not a matrix or data frame
}
\description{
Gower_Mat_Check
}
\keyword{internal}
