% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_checks.R
\name{K_Range_Check}
\alias{K_Range_Check}
\title{K_Range_Check}
\usage{
K_Range_Check(silhouette_kmin, silhouette_kmax)
}
\arguments{
\item{silhouette_kmin}{Numeric value indicating the minimum k value to be included in the silhouette plot.}

\item{silhouette_kmax}{Numeric value indicating the maximum k value to be included in the silhouette plot.}
}
\value{
Returns an error message if silhouette_kmin and/or silhouette_kmax are incompatible.
}
\description{
K_Range_Check
}
\keyword{internal}
