% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_checks.R
\name{Method_Check}
\alias{Method_Check}
\title{Method_Check}
\usage{
Method_Check(method)
}
\arguments{
\item{method}{A string indicating the agglomerative method to be used in hierarchical clustering.}
}
\value{
Returns an error message if method is incompatible.
}
\description{
Method_Check
}
\keyword{internal}
