% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower_dist_mat.R
\name{Gower_Cluster}
\alias{Gower_Cluster}
\title{Gower_Cluster}
\usage{
Gower_Cluster(data.x, var.type.vec, var.weight.vec = NULL, dend = FALSE)
}
\arguments{
\item{data.x}{n x p dataframe object (allows for mixed data types) or a matrix object (single data type).}

\item{var.type.vec}{p x 1 vector with elements (values of 0-3) indicating the intended data type of each column of data.x. Mode numeric (0) will be considered as interval scaled variables; mode character or class factor (1) will be considered as categorical nominal variables; class ordered (2) will be considered as categorical ordinal variables; mode logical (3) will be considered as binary asymmetric variables (view details of gower.dist() for more info).}

\item{var.weight.vec}{(Optional) p x 1 vector with elements indicating variable weights applied in Gower's Distance calculations. User can specify different weights for different variables by providing a numeric value for each variable contributing to the distance. Length should equal the number of variables considered in calculating distance. Entered weights are scaled to sum up to 1.}

\item{dend}{(Optional) boolean input, "TRUE" indicating dendrogram plot is a desired output, "FALSE" otherwise.}
}
\value{
A matrix
\item{gower.mat}{ n x n matrix representing the pairwise Gower's Distances between observations of data.x }
\item{dend.plot}{ (Optional) dendgrogram plot indicating potential clusters of data.x observations based on pairwise Gower's Distances }
}
\description{
Gower_Cluster
}
\examples{

##### Example 1

#####################
##### Example 2

}
