% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_checks.R
\name{Var_Type_Check}
\alias{Var_Type_Check}
\title{Var_Type_Check}
\usage{
Var_Type_Check(var.type.vec)
}
\arguments{
\item{var.type.vec}{}
}
\value{
Returns an error message if var.type.vec does not follow one or more of the requirements mentioned below.
}
\description{
Var_Type_Check
}
\keyword{internal}
