% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_checks.R
\name{Var_Weight_Check}
\alias{Var_Weight_Check}
\title{Var_Weight_Check}
\usage{
Var_Weight_Check(var.weight.vec)
}
\arguments{
\item{var.weight.vec}{}
}
\value{
Returns an error message if var.weight.vec does not follow one or more of the requirements mentioned below.
}
\description{
Var_Weight_Check
}
\keyword{internal}
